backend/requirements.txt
# Core Web Framework
Flask==3.0.3
flask-cors==4.0.1

# HTTP & API
requests==2.32.3
feedparser==6.0.11

# NLP & Embeddings
sentence-transformers==3.0.1
transformers==4.44.2
torch>=2.0.0

# Machine Learning / Analytics
scikit-learn==1.5.1
numpy==1.26.4
scipy==1.13.1

# Vector Store
faiss-cpu==1.8.0.post1

# Database
psycopg2-binary==2.9.9

# LLM API
openai==1.42.0

# Utilities
python-dotenv==1.0.1

# Optional (better performance / future use)
tqdm==4.66.5
